# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FinFlow API",
    "version": "1.0.0",
    "description": "All endpoints require an authenticated session (Flask-Login cookie, set by /login) and are scoped to the logged-in user's own accounts. State-changing requests (POST/PATCH/DELETE) also require a CSRF token, sent as header `X-CSRFToken` (get one from the `csrf-token` <meta> tag on any page, or the `csrf_token` hidden field on the login/signup forms)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "X-CSRFToken",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Required on POST/PATCH/DELETE."
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/transactions": {
      "get": {
        "summary": "Paginated, searchable, filterable transaction list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 200
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Matches merchant name."
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DEBIT",
                "CREDIT"
              ]
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "amount_min",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "amount_max",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of the caller's own transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "page_size": {
                      "type": "integer"
                    },
                    "transactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "transaction_id": {
                            "type": "string"
                          },
                          "account_id": {
                            "type": "string"
                          },
                          "merchant_id": {
                            "type": "string"
                          },
                          "category_id": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "transaction_type": {
                            "type": "string",
                            "enum": [
                              "DEBIT",
                              "CREDIT"
                            ]
                          },
                          "transaction_ts": {
                            "type": "string"
                          },
                          "is_fraud": {
                            "type": "boolean",
                            "description": "Ground-truth fraud label (only ever set on generated data; used to evaluate the models). Never written by an import or the stream."
                          },
                          "is_recurring": {
                            "type": "boolean"
                          },
                          "fraud_score": {
                            "type": "number",
                            "nullable": true,
                            "minimum": 0,
                            "maximum": 1,
                            "description": "Calibrated 0-1 anomaly probability from the fraud model, set when the row was imported or streamed; null if unscored."
                          },
                          "is_flagged": {
                            "type": "boolean",
                            "description": "True if fraud_score >= the user's alert sensitivity (the same rule the Alerts page uses) or the row carries a ground-truth label."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter value"
          }
        }
      }
    },
    "/transactions/{transaction_id}/category": {
      "patch": {
        "summary": "Edit a transaction's category, optionally creating a standing rule",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "category_id"
                ],
                "properties": {
                  "category_id": {
                    "type": "string"
                  },
                  "create_rule": {
                    "type": "boolean",
                    "default": false,
                    "description": "If true, saves a standing rule for this merchant (for a UPI payment, the payee), re-categorizes all of this user's existing transactions with the same merchant identity, and applies the rule to future uploads."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction_id": {
                      "type": "string"
                    },
                    "category_id": {
                      "type": "string"
                    },
                    "rule_created": {
                      "type": "boolean"
                    },
                    "transactions_updated": {
                      "type": "integer",
                      "description": "How many transactions changed category (1 unless a rule was created)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown category"
          },
          "404": {
            "description": "Transaction not found or not owned by this user"
          }
        }
      }
    },
...
  }
}